+++ src/test/java/today/sesac/shoutify/global/advice/GlobalExceptionHandlerTest.java
@@ -1,8 +1,13 @@
 package today.sesac.shoutify.global.advice;
 
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
+import jakarta.validation.Valid;
+import jakarta.validation.constraints.Email;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,12 +22,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-
-import jakarta.validation.Valid;
-import jakarta.validation.constraints.Email;
-import today.sesac.shoutify.auth.config.TestSecurityConfig;
+import today.sesac.shoutify.global.config.TestSecurityConfig;
 import today.sesac.shoutify.global.exception.LoginRequiredException;
 import today.sesac.shoutify.global.exception.PermissionRequiredException;
 
@@ -33,126 +33,129 @@ import today.sesac.shoutify.global.exception.PermissionRequiredException;
 @WebMvcTest(controllers = GlobalExceptionHandlerTest.class)
 class GlobalExceptionHandlerTest {
 
-	static final String IS_SUCCESS = "$.isSuccess";
-	static final String ERROR_NAME = "$.error.name";
-	static final String ERROR_MESSAGE = "$.error.message";
-	static final String ERROR_PARAM = "$.error.param";
+    static final String IS_SUCCESS = "$.isSuccess";
+    static final String ERROR_NAME = "$.error.name";
+    static final String ERROR_MESSAGE = "$.error.message";
+    static final String ERROR_PARAM = "$.error.param";
 
-	@Autowired
-	private MockMvc mockMvc;
+    @Autowired
+    private MockMvc mockMvc;
 
-	@Test
-	@DisplayName("[예외]-RuntimeException 처리 테스트")
-	void handleRuntimeException() throws Exception {
-		mockMvc.perform(get("/test/error/runtime"))
-			.andExpect(status().isInternalServerError())
-			.andExpect(jsonPath(IS_SUCCESS).value(false))
-			.andExpect(jsonPath(ERROR_NAME).value("INTERNAL_SERVER"))
-			.andExpect(jsonPath(ERROR_MESSAGE).value("알 수 없는 문제가 발생했습니다. 잠시 후 다시 시도해 주세요."))
-			.andExpect(jsonPath(ERROR_PARAM).isEmpty())
-			.andDo(MockMvcResultHandlers.print());
-	}
+    @Test
+    @DisplayName("[예외]-RuntimeException 처리 테스트")
+    void handleRuntimeException() throws Exception {
+        mockMvc.perform(get("/test/error/runtime"))
+                .andExpect(status().isInternalServerError())
+                .andExpect(jsonPath(IS_SUCCESS).value(false))
+                .andExpect(jsonPath(ERROR_NAME).value("INTERNAL_SERVER"))
+                .andExpect(jsonPath(ERROR_MESSAGE).value("알 수 없는 문제가 발생했습니다. 잠시 후 다시 시도해 주세요."))
+                .andExpect(jsonPath(ERROR_PARAM).isEmpty())
+                .andDo(MockMvcResultHandlers.print());
+    }
 
-	@Test
-	@DisplayName("[예외]-LoginRequiredException 처리 테스트")
-	void handleLoginRequiredException() throws Exception {
-		mockMvc.perform(get("/test/error/login-required"))
-			.andExpect(status().isUnauthorized())
-			.andExpect(jsonPath(IS_SUCCESS).value(false))
-			.andExpect(jsonPath(ERROR_NAME).value("LOGIN_REQUIRED"))
-			.andExpect(jsonPath(ERROR_MESSAGE).value("로그인이 필요합니다."))
-			.andExpect(jsonPath(ERROR_PARAM).value("member"))
-			.andDo(MockMvcResultHandlers.print());
-	}
+    @Test
+    @DisplayName("[예외]-LoginRequiredException 처리 테스트")
+    void handleLoginRequiredException() throws Exception {
+        mockMvc.perform(get("/test/error/login-required"))
+                .andExpect(status().isUnauthorized())
+                .andExpect(jsonPath(IS_SUCCESS).value(false))
+                .andExpect(jsonPath(ERROR_NAME).value("LOGIN_REQUIRED"))
+                .andExpect(jsonPath(ERROR_MESSAGE).value("로그인이 필요합니다."))
+                .andExpect(jsonPath(ERROR_PARAM).value("member"))
+                .andDo(MockMvcResultHandlers.print());
+    }
 
-	@Test
-	@DisplayName("[예외]-PermissionRequiredException 처리 테스트")
-	void handlePermissionRequiredException() throws Exception {
-		mockMvc.perform(get("/test/no-permission"))
-			.andExpect(status().isForbidden())
-			.andExpect(jsonPath(IS_SUCCESS).value(false))
-			.andExpect(jsonPath(ERROR_NAME).value("PERMISSION_REQUIRED"))
-			.andExpect(jsonPath(ERROR_MESSAGE).value("권한이 없습니다. 다시 확인해 주세요."))
-			.andExpect(jsonPath(ERROR_PARAM).value("member"))
-			.andDo(MockMvcResultHandlers.print());
-	}
+    @Test
+    @DisplayName("[예외]-PermissionRequiredException 처리 테스트")
+    void handlePermissionRequiredException() throws Exception {
+        mockMvc.perform(get("/test/no-permission"))
+                .andExpect(status().isForbidden())
+                .andExpect(jsonPath(IS_SUCCESS).value(false))
+                .andExpect(jsonPath(ERROR_NAME).value("PERMISSION_REQUIRED"))
+                .andExpect(jsonPath(ERROR_MESSAGE).value("권한이 없습니다. 다시 확인해 주세요."))
+                .andExpect(jsonPath(ERROR_PARAM).value("member"))
+                .andDo(MockMvcResultHandlers.print());
+    }
 
-	@Test
-	@DisplayName("[예외]-NoHandlerFoundException 처리 테스트")
-	void handleNoHandlerFoundException() throws Exception {
-		mockMvc.perform(get("/test/not-found"))
-			.andExpect(status().isNotFound())
-			.andExpect(jsonPath(IS_SUCCESS).value(false))
-			.andExpect(jsonPath(ERROR_NAME).value("UNDISCOVERED"))
-			.andExpect(jsonPath(ERROR_MESSAGE).value("요청하신 정보를 찾을 수 없습니다."))
-			.andExpect(jsonPath(ERROR_PARAM).isEmpty());
-	}
+    @Test
+    @DisplayName("[예외]-NoHandlerFoundException 처리 테스트")
+    void handleNoHandlerFoundException() throws Exception {
+        mockMvc.perform(get("/test/not-found"))
+                .andExpect(status().isNotFound())
+                .andExpect(jsonPath(IS_SUCCESS).value(false))
+                .andExpect(jsonPath(ERROR_NAME).value("UNDISCOVERED"))
+                .andExpect(jsonPath(ERROR_MESSAGE).value("요청하신 정보를 찾을 수 없습니다."))
+                .andExpect(jsonPath(ERROR_PARAM).isEmpty());
+    }
 
-	@Test
-	@DisplayName("[예외]-MethodArgumentNotValidException 처리 테스트")
-	void handleMethodArgumentNotValidException() throws Exception {
+    @Test
+    @DisplayName("[예외]-MethodArgumentNotValidException 처리 테스트")
+    void handleMethodArgumentNotValidException() throws Exception {
 
-		ObjectMapper objectMapper = new ObjectMapper();
+        ObjectMapper objectMapper = new ObjectMapper();
 
-		String invalidJson = objectMapper.writeValueAsString(new TestReq("invalidEmailFormat"));
+        String invalidJson = objectMapper.writeValueAsString(new TestReq("invalidEmailFormat"));
 
-		mockMvc.perform(post("/test/valid")
-				.contentType(MediaType.APPLICATION_JSON)
-				.content(invalidJson))
-			.andExpect(status().isBadRequest())
-			.andExpect(jsonPath(IS_SUCCESS).value(false))
-			.andExpect(jsonPath("$.error[0].name").value("INVALID_REQUEST"))
-			.andExpect(jsonPath("$.error[0].message").value("유효하지 않은 형식입니다. 요청 값을 다시 확인해 주세요."))
-			.andExpect(jsonPath("$.error[0].param").value("email"))
-			.andDo(MockMvcResultHandlers.print());
-	}
+        mockMvc.perform(post("/test/valid")
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .content(invalidJson))
+                .andExpect(status().isBadRequest())
+                .andExpect(jsonPath(IS_SUCCESS).value(false))
+                .andExpect(jsonPath("$.error[0].name").value("INVALID_REQUEST"))
+                .andExpect(jsonPath("$.error[0].message").value("유효하지 않은 형식입니다. 요청 값을 다시 확인해 주세요."))
+                .andExpect(jsonPath("$.error[0].param").value("email"))
+                .andDo(MockMvcResultHandlers.print());
+    }
 
-	/**
-	 * 테스트용 컨트롤러
-	 */
-	@RestController
-	static class TestController {
+    /**
+     * 테스트용 컨트롤러
+     */
+    @RestController
+    static class TestController {
 
-		@GetMapping("/test/error/runtime")
-		public void testEndpoint() {
-			throw new RuntimeException("Test Runtime Exception");
-		}
+        @GetMapping("/test/error/runtime")
+        public void testEndpoint() {
+            throw new RuntimeException("Test Runtime Exception");
+        }
 
-		@GetMapping("/test/error/login-required")
-		public void loginRequiredEndpoint() {
-			throw new LoginRequiredException("member", "로그인이 필요합니다. - stack trace에 기록된 메시지");
-		}
+        @GetMapping("/test/error/login-required")
+        public void loginRequiredEndpoint() {
+            throw new LoginRequiredException("member", "로그인이 필요합니다. - stack trace에 기록된 메시지");
+        }
 
-		@GetMapping("/test/no-permission")
-		public void noPermissionEndpoint() {
-			throw new PermissionRequiredException("member", "권한이 없습니다.");
-		}
+        @GetMapping("/test/no-permission")
+        public void noPermissionEndpoint() {
+            throw new PermissionRequiredException("member", "권한이 없습니다.");
+        }
 
-		@PostMapping("/test/valid")
-		public void validTest(@Valid @RequestBody TestReq request) {
-			// 요청이 유효하면 아무 일도 하지 않음
-		}
+        @PostMapping("/test/valid")
+        public void validTest(@Valid @RequestBody TestReq request) {
+            // 요청이 유효하면 아무 일도 하지 않음
+        }
 
-	}
+    }
 
-	/**
-	 * 테스트용 요청 DTO
-	 * @param email 이메일
-	 */
-	record TestReq(
-		@Email(message = "유효하지 않은 형식입니다. 요청 값을 다시 확인해 주세요.")
-		String email
-	) {
-	}
+    /**
+     * 테스트용 요청 DTO
+     *
+     * @param email 이메일
+     */
+    record TestReq(
+            @Email(message = "유효하지 않은 형식입니다. 요청 값을 다시 확인해 주세요.")
+            String email
+    ) {
 
-	/**
-	 * 테스트용 컨트롤러를 빈으로 등록하는 설정 클래스
-	 */
-	@TestConfiguration
-	static class TestConfig {
-		@Bean
-		public TestController testController() {
-			return new TestController();
-		}
-	}
+    }
+
+    /**
+     * 테스트용 컨트롤러를 빈으로 등록하는 설정 클래스
+     */
+    @TestConfiguration
+    static class TestConfig {
+
+        @Bean
+        public TestController testController() {
+            return new TestController();
+        }
+    }
 }
